refactor(task-tracker): relocate membership store under the module; guard the kernel boundary
