+++ src/modules/task-tracker/membership-store.ts
@@ -4,12 +4,12 @@
 // See LICENSE in the project root for details.
 
 import type { MembershipStore, UserLabelResolver } from '../../ports/membership-store.js'
-import { runMembershipBackfill } from '../../providers/membership/backfill.js'
 import {
   defaultMembershipDeps,
   ensureWorkspaceMember,
   markMemberInactive,
-} from '../../providers/membership/ensure-member.js'
+  runMembershipBackfill,
+} from './membership/index.js'
 
 /** The task-tracker adapter binding the membership store implementation to the port. */
 export const taskProviderMembershipStore: MembershipStore = {
